rpi: expose the firmware provided FDT blob in ${fdt_addr}
If the fw_dtb_pointer saved at an early stage points to a valid FDT
blob, let's expose it in ${fdt_addr} from where it can be used in
many different ways.

For example, the following script will extract the firmware provided
cmdline from the tree, then hand over the blob to a previously loaded
kernel:

if fdt addr ${fdt_addr}
then
	fdt get value bootargs /chosen bootargs
	bootz ${kernel_addr_r} - ${fdt_addr}
fi

Alternately, when relying on an extlinux.conf file, it's enough to
omit any FDT statement to make U-Boot automagically pick ${fdt_addr}.

Please note that for this to work it is necessary to tag the U-Boot
binary with a recent version of the mkknlimg script found in the
Rasperry Fundation kernel tree:

<kernel>/scripts/mkknlimg --dtok <uboot>/u-boot.bin /boot/u-boot.bin

Series-to: u-boot
Series-version: 2
Series-changes: 2
- add the extern declaration in rpi.c
- fix 64-bits compatibility issues
- rename the environment variable to fdt_addr
- document possible uses of fdt_addr

Cover-letter:
rpi: enable passthrough of FDT blob provided by the firmware
Raspberry firmware used to pass a FDT blob at a fixed address (0x100),
but this is not true anymore. The address now depends on both the
memory size and the blob size [1].

If one wants to passthrough this FDT blob to the kernel, the most
reliable way is to save its address from the r2/x0 register passed to
U-Boot in the entry point and expose it in a environment variable for
further processing.

[1] https://www.raspberrypi.org/forums//viewtopic.php?f=107&t=134018
END